Upgraded iex-cloud to v3 framework
