+++ .changeset/brown-snails-stare.md
@@ -1,0 +1,5 @@
+---
+'@chainlink/iex-cloud-adapter': major
+---
+
+Upgraded adapter from v2 to v3 framework
