+++ packages/core/legos/package.json
@@ -121,7 +121,6 @@
     "@chainlink/gsr-adapter": "workspace:*",
     "@chainlink/icap-adapter": "workspace:*",
     "@chainlink/iex-cloud-adapter": "workspace:*",
-    "@chainlink/iex-cloud-test-adapter": "workspace:*",
     "@chainlink/intrinio-adapter": "workspace:*",
     "@chainlink/intrinio-test-adapter": "workspace:*",
     "@chainlink/ipfs-adapter": "workspace:*",
