+++ packages/core/legos/src/sources.ts
@@ -90,7 +90,6 @@ import * as graphql from '@chainlink/graphql-adapter'
 import * as gsr from '@chainlink/gsr-adapter'
 import * as icap from '@chainlink/icap-adapter'
 import * as iex_adapter from '@chainlink/iex-cloud-adapter'
-import * as iex_cloud_test from '@chainlink/iex-cloud-test-adapter'
 import * as intrinio from '@chainlink/intrinio-adapter'
 import * as intrinio_test from '@chainlink/intrinio-test-adapter'
 import * as ipfs from '@chainlink/ipfs-adapter'
@@ -256,7 +255,6 @@ export default {
   gsr,
   icap,
   iex_adapter,
-  iex_cloud_test,
   intrinio,
   intrinio_test,
   ipfs,
