+++ packages/core/legos/tsconfig.json
@@ -287,9 +287,6 @@
       "path": "../../sources/iex-cloud"
     },
     {
-      "path": "../../sources/iex-cloud-test"
-    },
-    {
       "path": "../../sources/intrinio"
     },
     {
