--- packages/sources/iex-cloud-test/CHANGELOG.md
@@ -1,25 +1,0 @@
-# @chainlink/iex-cloud-test-adapter
-
-## 1.0.3
-
-### Patch Changes
-
-- 77ad946a9: Bumped framework version
-
-## 1.0.2
-
-### Patch Changes
-
-- 4245b4d8b: Bumped framework version
-
-## 1.0.1
-
-### Patch Changes
-
-- 8fe5be225: Bumped framework version
-
-## 1.0.0
-
-### Major Changes
-
-- a8b11e3a7: Initial version of the adapter
