--- packages/sources/iex-cloud-test/src/endpoint/index.ts
@@ -1,3 +1,0 @@
-export { endpoint as crypto } from './crypto'
-export { endpoint as eod } from './eod'
-export { endpoint as stock } from './stock'
