--- packages/sources/iex-cloud-test/test-payload.json
@@ -1,6 +1,0 @@
-{
- "requests": [{
-  "base": "IWM",
-  "endpoint": "stock"
- }]
-}
