--- packages/sources/iex-cloud/schemas/env.json
@@ -1,20 +1,0 @@
-{
-  "$id": "https://external-adapters.chainlinklabs.com/schemas/iex-cloud-adapter.json",
-  "title": "Chainlink External Adapter for IEX Cloud",
-  "required": ["API_KEY"],
-  "type": "object",
-  "properties": {
-    "API_KEY": {
-      "type": "string"
-    },
-    "API_ENDPOINT": {
-      "default": "https://cloud.iexapis.com/stable",
-      "type": "string"
-    }
-  },
-  "allOf": [
-    {
-      "$ref": "https://external-adapters.chainlinklabs.com/schemas/ea-bootstrap.json"
-    }
-  ]
-}
