+++ packages/sources/iex-cloud/src/endpoint/index.ts
@@ -1,9 +1,3 @@
-import type { TInputParameters as StockInputParameters } from './stock'
-import type { TInputParameters as CryptoInputParameters } from './crypto'
-import type { TInputParameters as EodInputParameters } from './eod'
-
-export type TInputParameters = StockInputParameters | CryptoInputParameters | EodInputParameters
-
-export * as stock from './stock'
-export * as crypto from './crypto'
-export * as eod from './eod'
+export { endpoint as crypto } from './crypto'
+export { endpoint as eod } from './eod'
+export { endpoint as stock } from './stock'
