+++ packages/sources/iex-cloud/src/endpoint/stock.ts
@@ -1,39 +1,23 @@
-import { Requester, util, Validator } from '@chainlink/ea-bootstrap'
-import type { ExecuteWithConfig, Config, InputParameters } from '@chainlink/ea-bootstrap'
-import { NAME as AdapterName } from '../config'
-import { ResponseSchema } from './eod'
+import { HttpTransport } from '@chainlink/external-adapter-framework/transports'
+import { AdapterEndpoint } from '@chainlink/external-adapter-framework/adapter'
+import {
+  buildHttpRequestBody,
+  constructEntries,
+  inputParameters,
+  StockEndpointTypes,
+} from '../stock-utils'
 
-export const supportedEndpoints = ['stock']
-
-export type TInputParameters = { base: string }
-export const inputParameters: InputParameters<TInputParameters> = {
-  base: {
-    aliases: ['from', 'coin', 'asset', 'symbol'],
-    description: 'The symbol to query',
-    required: true,
-    type: 'string',
+export const httpTransport = new HttpTransport<StockEndpointTypes>({
+  prepareRequests: (params, config) => {
+    return buildHttpRequestBody(params, config)
   },
-}
+  parseResponse: (params, res) => {
+    return constructEntries(res.data, params, 'latestPrice')
+  },
+})
 
-export const execute: ExecuteWithConfig<Config> = async (request, _, config) => {
-  const validator = new Validator(request, inputParameters)
-
-  const jobRunID = validator.validated.id
-  const base = validator.overrideSymbol(AdapterName, validator.validated.data.base)
-  const url = util.buildUrlPath('stock/:base/quote', { base: base.toUpperCase() })
-
-  const params = {
-    token: config.apiKey,
-  }
-
-  const reqConfig = {
-    ...config.api,
-    params,
-    url,
-  }
-
-  const response = await Requester.request<ResponseSchema>(reqConfig)
-  const result = Requester.validateResultNumber(response.data, ['latestPrice'])
-
-  return Requester.success(jobRunID, Requester.withResult(response, result), config.verbose)
-}
+export const endpoint = new AdapterEndpoint<StockEndpointTypes>({
+  name: 'stock',
+  transport: httpTransport,
+  inputParameters: inputParameters,
+})
