+++ packages/sources/iex-cloud/src/index.ts
@@ -1,9 +1,25 @@
-import { expose } from '@chainlink/ea-bootstrap'
-import { makeExecute, endpointSelector } from './adapter'
-import { makeConfig, NAME } from './config'
-import rateLimit from './config/limits.json'
+import { expose, ServerInstance } from '@chainlink/external-adapter-framework'
+import { PriceAdapter } from '@chainlink/external-adapter-framework/adapter'
+import { config } from './config'
+import { crypto, eod, stock } from './endpoint'
 
-const adapterContext = { name: NAME, rateLimit }
+export const adapter = new PriceAdapter({
+  defaultEndpoint: stock.name,
+  name: 'IEXCLOUD',
+  config,
+  endpoints: [crypto, stock, eod],
+  rateLimiting: {
+    tiers: {
+      individual: {
+        rateLimit1h: 6944.44444444,
+        note: 'only mentions monthly limit',
+      },
+      business: {
+        rateLimit1h: 208333.333333,
+        note: 'only mentions monthly limit',
+      },
+    },
+  },
+})
 
-const { server } = expose(adapterContext, makeExecute(), undefined, endpointSelector)
-export { NAME, makeExecute, makeConfig, server }
+export const server = (): Promise<ServerInstance | undefined> => expose(adapter)
