--- packages/sources/iex-cloud/test/unit/adapter.test.ts
@@ -1,40 +1,0 @@
-import { AdapterError, Requester } from '@chainlink/ea-bootstrap'
-import { assertError } from '@chainlink/ea-test-helpers'
-import { AdapterRequest } from '@chainlink/ea-bootstrap'
-import { makeExecute } from '../../src/adapter'
-import { TInputParameters } from '../../src/endpoint'
-
-describe('execute', () => {
-  const jobID = '1'
-  const execute = makeExecute()
-
-  describe('validation error', () => {
-    const requests = [
-      { name: 'empty body', testData: {} },
-      { name: 'empty data', testData: { data: {} } },
-      {
-        name: 'base not supplied on crypto',
-        testData: { id: jobID, data: { quote: 'USD', endpoint: 'crypto' } },
-      },
-      {
-        name: 'quote not supplied on crypto',
-        testData: { id: jobID, data: { base: 'ETH', endpoint: 'crypto' } },
-      },
-      {
-        name: 'symbol not supplied on stock',
-        testData: { id: jobID, data: { quote: 'USD' } },
-      },
-    ]
-
-    requests.forEach((req) => {
-      it(`${req.name}`, async () => {
-        try {
-          await execute(req.testData as AdapterRequest<TInputParameters>, {})
-        } catch (error) {
-          const errorResp = Requester.errored(jobID, error as AdapterError)
-          assertError({ expected: 400, actual: errorResp.statusCode }, errorResp, jobID)
-        }
-      })
-    })
-  })
-})
